requirements.txt
fastapi==0.115.10
uvicorn==0.27.1
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.1
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
python-multipart==0.0.6
sqlalchemy==2.0.23
aiosqlite==0.19.0
openai==1.3.0 